Refactor routes Added controller directory to handle route callback functionality. Added routes directory to add controller callbacks to router object Simplified winston transport add
