+++ middlewares/config.js
@@ -1,10 +1,11 @@
+const winston = require('winston');
 const bodyParser = require('body-parser');
 
-module.exports = (app, logger) => {
+module.exports = app => {
   app.use(bodyParser.json());
   app.use(
     require('morgan')('combined', {
-      stream: logger.stream
+      stream: winston.stream
     })
   );
 };
